Moves NetCDF4ArrayConverter & NetCDFDomainConverter into new module
Editing converter.py was inconvenient due to length of file.
